rxhub-provider-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
httpx==0.27.2
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
python-multipart==0.0.12
slowapi==0.1.9
redis==5.2.1
